Moves language validation methods to ScopeInfo
Fixes: SIRI-368
