+++ src/main/java/sirius/web/http/LangHelper.java
@@ -46,7 +46,7 @@ public class LangHelper {
             if (m.matches()) {
                 double q = Value.of(m.group(4)).asDouble(1.0d);
                 String language = m.group(1);
-                if (q > bestQ && UserContext.get().getUserManager().isSupportedLanguage(language)) {
+                if (q > bestQ && UserContext.get().getScope().isSupportedLanguage(language)) {
                     bestQ = q;
                     currentLang = language;
                 }
